+++ lib/src/types/map_animation.dart
@@ -1,22 +1,37 @@
 part of yandex_mapkit;
 
+/// The animation that is used to switch between states.
 class MapAnimation extends Equatable {
   const MapAnimation({
-    this.smooth = true,
-    this.duration = kAnimationDuration
+    this.type = MapAnimationType.smooth,
+    this.duration = 2.0
   });
 
+  /// Animation duration, in seconds.
   final double duration;
-  final bool smooth;
 
-  static const double kAnimationDuration = 2.0;
+  /// Smooth interpolation between start and finish states or movement with constant speed during animation time.
+  final MapAnimationType type;
 
   @override
   List<Object> get props => <Object>[
-    smooth,
+    type,
     duration
   ];
 
   @override
   bool get stringify => true;
+
+  Map<String, dynamic> toJson() {
+    return <String, dynamic>{
+      'type': type.index,
+      'duration': duration
+    };
+  }
+}
+
+/// Animation types
+enum MapAnimationType {
+  smooth,
+  linear
 }